+++ icons/Icon/IconProps.ts
@@ -1,3 +1,22 @@
+import { SVGProps } from 'react';
+
+export interface InternalIconProps
+  extends Omit<SVGProps<SVGSVGElement>, 'width' | 'height' | 'className' | 'height' | 'children'> {
+  // TODO: update type to "typeof colors[keyof typeof colors]" when colors are defined
+  fill?: string;
+  /**
+   * Give the icon a description. It will show up in a tooltip after hovering over the icon for 3s.
+   * This is a default browser feature and very helpful to users to understand what the icon does.
+   */
+  title?: string;
+  /**
+   * If the icon is labelled by an independent element, then provide that element's id. This is
+   * used to set the aria-labelledby value on the icon svg. If the `title` prop is also provided,
+   * then `titleId` will be used to set the id on the `<title>` element.
+   */
+  titleId?: string;
+}
+
 export type IconProps = {
   /** Identifies the icon */
   kind: IconKind;
@@ -5,13 +24,12 @@ export type IconProps = {
    * Specifies the icon size. The actual icon size is determined based on this prop and the
    * current scale.
    */
-
   size: IconSize;
   /** Set the fill color of the icon */
   // TODO: update type to allowed colors
-
   fill?: string;
-};
+} & InternalIconProps;
+
 export type IconSize = 'L' | 'M' | 'S' | 'XS';
 export type Pixels = 32 | 24 | 16 | 12 | 8;
 type IconKind =